packages/knowledge_base/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-academy-knowledge-base"
version = "0.1.0"
description = "Knowledge base indexing and MCP service"
requires-python = "==3.11.14"
dependencies = [
  "ai-academy-shared-config",
  "fastapi[standard]",
  "faster-whisper",
  "httpx",
  "langchain-community",
  "langchain-text-splitters",
  "mcp",
  "pydantic-settings",
  "pypdf",
  "qdrant-client",
  "rich",
]

[tool.uv.sources]
"ai-academy-shared-config" = { workspace = true }

[tool.setuptools]
package-dir = { "" = "src" }
py-modules = ["main", "mcp_tools"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["indexer*", "services*"]
